requirements.txt
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
discord-protos==0.0.2
discord.py-self @ git+https://github.com/dolfies/discord.py-self.git
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pytz==2025.1
multidict==6.1.0
protobuf==5.29.3
Pygments==2.19.1
requests==2.32.3
rich==13.9.4
tzlocal==5.3
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
pillow==10.2.0
playsound3==0.5.2
plyer==2.1.0
psutil==5.9.8
python-dotenv==1.0.1
